obj_id von int auf long geandert
git-svn-id: https://svn.r7r.at/sepm/trunk@326 240b89b8-79f3-4156-a269-5c34a98ffd57

+++ website/src/main/java/at/fakeroot/sepm/client/serialize/ObjectSearchService.java
@@ -12,6 +12,6 @@ public interface ObjectSearchService extends RemoteService
 {
 	public SearchResult search(BoundingBox box, String what);
 	
-	public ObjectDetails getDetailHTML(int objId);
+	public ObjectDetails getDetailHTML(long objId);
 }
